pyproject.toml
[project]
name = "api-agentai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "flask>=3.1.2",
    "google-adk>=1.19.0",
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
]
